Install Sublime Text and symlink correct paths
This installation depends on the file `sublime.dmg` being present in
the `~/Dropbox/dotfiles/chef/cookbooks/sublime-text/files/default/`
path. This file is not included in this repository. When Sublime Text 2
1.0 (or 2.0?) is released, this will be changed to fetch the file from
their servers.

If a license file is present in `~/Dropbox/dotfiles/Application
Support/Sublime Text 2/Config`, it will be symlinked to your
Application Support folder.