Clean up bootstrap.sh and use path variables
You can now use the script with an argument which will be the location
where your dotfiles will live. If no argument is given `~/dotfiles` is
assumed.

Also, the script now gets the Dropbox directory from a
get_dropbox_folder.sh utility which checks the dropbox sqlite3 database
for the correct path.