Modify the puppet gem because we need(ed?) the HOME variable in Homebrew
After a lengthy discussion on the #homebrew chat, it turns out Homebrew errors out when no HOME path is available. But Puppet strips this environment variable because of unclear reasons (see: http://grokbase.com/t/gg/puppet-users/12a35t566x/puppet-3-killed-my-environment-variables).

For now, we will have to make due with modifying the Puppet gem to allow the HOME variable to pass through from puppet to Homebrew. In the chat it was indicated that this will eventually be solved either in Homebrew or in Puppet.