chore: add noop bootstrap script
The purpose of this script will be to bootstrap macOS from start to
finish. The script is intended to run frequently in order to keep the
system in pristine condition with the latest software installed and
patches applied.

The script should not perform any (undocumented) destructive actions, as
it will also run when the system has been running for weeks or months.